refactor(ImageDetection): extract AnnotationPopup component
